requirements.txt
boto3>=1.18.46
convertapi>=1.4.0
coverage>=6.1.2
Django>=3.0,<4.0
djangorestframework>=3.12.4
django-csp>=3.7
django-filter>=21.1
django-mailjet>=0.3.1
django-storages>=1.11.1
djhtml>=1.4.9
dj-database-url>=0.5.0
dj-static>=0.0.6
docxtpl>=0.12.0
drf-yasg>=1.20.0
gunicorn>=20.1.0
openpyxl>=3.0.7
Pillow>=8.3.2
pre-commit>=2.13.0
psycopg2-binary>=2.8
pylint>=2.9.5
python-decouple>=3.4
python-docx>=0.8.11
jinja2>=3.0.2
whitenoise>=5.3.0
